Shooting Revamp
Mainly worked on the revamp of the shooting system to switch to controller input as this now a high priority. Basic 8 way aiming is in so just need to implement the other shooters logic.

Other misc tasks such as shop updates, difficulty selection (whihc doesnt actually do anything at the moment) and other cleanup work.